[tests] Refactored test cases
