+++ tests/integration/workflows.py
@@ -22,99 +22,16 @@ from pycti.connector.v2.libs.orchestrator_schemas import (
     ConnectorCreate,
     Connector,
     ConfigCreate,
-    Config
+    Config,
 )
 from app.extensions import elastic
-
-
-def create_connector_and_config_stix(test_client) -> str:
-    # Register Connector
-    stix_worker = ConnectorCreate(
-        uuid="7af04c62-5783-497b-a49b-baba88b8802d",
-        name="StixWorker",
-        type="STIX_IMPORT",
-        queue="stix_import",
-        config_schema=None
-    )
-    response = test_client.post("/connector/", json=stix_worker.dict())
-    stix_worker = json.loads(response.data.decode())
-    print(stix_worker)
-    assert response.status_code == 201, f"Error {response.data.decode()}"
-
-    stix_worker_connector = Connector(**stix_worker["connector"])
-
-    # Register Connector Config
-    stix_worker_config = ConfigCreate(
-        connector_id=stix_worker_connector.id,
-        name="StixWorker Config",
-        config={}
-    )
-    response = test_client.post("/config/", json=stix_worker_config.dict())
-    assert response.status_code == 201, f"Error {response.data.decode()}"
-    stix_config = Config(**json.loads(response.data.decode()))
-    return stix_config.id
-
-
-def create_connector_and_config_external_import(test_client) -> str:
-    class TestExternalImportRunConfig(BaseModel):
-        ip: str
-
-    # Register Connector
-    external_import = ConnectorCreate(
-        uuid="61464bca-a9d6-47e4-a674-d50eb5df1354",
-        name="ExternalImport",
-        type=ConnectorType.EXTERNAL_IMPORT.value,
-        queue="external_import",
-        config_schema=TestExternalImportRunConfig.schema_json()
-    )
-    response = test_client.post("/connector/", json=external_import.dict())
-    external_import = json.loads(response.data.decode())
-    print(external_import)
-    assert response.status_code == 201, f"Error {response.data.decode()}"
-    external_import_connector = Connector(**external_import["connector"])
-
-    # Register Connector Config
-    external_import_config = ConfigCreate(
-        connector_id=external_import_connector.id,
-        name="EI Import",
-        config=TestExternalImportRunConfig(ip="192.168.14.1"),
-    )
-    a = external_import_config.dict()
-    print(f"sending config {a}")
-    response = test_client.post("/config/", json=external_import_config.dict())
-    assert response.status_code == 201, f"Error {response.data.decode()}"
-    ei_config = Config(**json.loads(response.data.decode()))
-    return ei_config.id
-
-
-def create_triggered_workflow(configs: List, test_client) -> str:
-    # Create Workflow
-    workflow = WorkflowCreate(
-        name="Test Workflow Trigered",
-        jobs=configs,
-        execution_type=ExecutionTypeEnum.triggered,
-        execution_args="",
-        token="123441",
-    )
-    response = test_client.post("/workflow/", json=workflow.dict())
-    assert response.status_code == 201, f"Error {response.data.decode()}"
-    workflow_config = Workflow(**json.loads(response.data.decode()))
-    return workflow_config.id
-
-
-def create_scheduled_workflow(configs: List, test_client) -> str:
-    # Create Workflow
-    workflow = WorkflowCreate(
-        name="Test Workflow Scheduled",
-        jobs=configs,
-        execution_type=ExecutionTypeEnum.scheduled,
-        execution_args="6",
-        token="123441",
-    )
-    response = test_client.post("/workflow/", json=workflow.dict())
-    assert response.status_code == 201, f"Error {response.data.decode()}"
-    workflow_config = Workflow(**json.loads(response.data.decode()))
-    return workflow_config.id
+from tests.utils.connector_creators import (
+    create_connector_and_config_external_import,
+    create_connector_and_config_stix,
+    create_triggered_workflow,
+    update_status,
+    create_scheduled_workflow,
+)
 
 
 @pytest.fixture()
@@ -143,26 +60,18 @@ def clear_test_suite(test_client):
     # elastic.connection.refresh(index=INDEX_NAME)
 
 
-def update_status(
-    test_client, config_id: str, run_id: str, status: State, result: Result = None
-) -> int:
-    update_schema = RunUpdate(
-        command="job_status",
-        parameters={"config_id": config_id, "status": status, "result": result},
-    )
-    response = test_client.put(f"/run/{run_id}", json=update_schema.dict())
-    # print(response.data.decode())
-    return response.status_code
-
-
 def test_triggered_workflow(test_client, caplog, clear_test_suite):
     """
     GIVEN a Flask application configured for testing
     WHEN the '/login' page is requested (GET)
     THEN check the response is valid
     """
-    external_import_config_id = create_connector_and_config_external_import(test_client)
-    stix_config_id = create_connector_and_config_stix(test_client)
+    (
+        external_import_config_id,
+        external_import_connector_id,
+        _,
+    ) = create_connector_and_config_external_import(test_client)
+    stix_config_id, stix_connector_id, _ = create_connector_and_config_stix(test_client)
 
     workflow_id = create_triggered_workflow(
         [stix_config_id, external_import_config_id], test_client
@@ -184,7 +93,9 @@ def test_triggered_workflow(test_client, caplog, clear_test_suite):
     run_id = None
     for record in caplog.records:
         if "Received RunContainer" in record.message:
-            run_container = RunContainer(**json.loads(record.message.split("RunContainer: ")[1]))
+            run_container = RunContainer(
+                **json.loads(record.message.split("RunContainer: ")[1])
+            )
             run_id = run_container.run_id
 
     assert run_id is not None, "Could not find Run ID in log output"
@@ -218,6 +129,16 @@ def test_triggered_workflow(test_client, caplog, clear_test_suite):
     assert run.status == State.finished.value
     assert run.result == Result.success.value
 
+    response = test_client.delete(f"/config/{external_import_config_id}")
+    assert response.status_code == 204
+    response = test_client.delete(f"/config/{stix_config_id}")
+    assert response.status_code == 204
+
+    response = test_client.delete(f"/connector/{external_import_connector_id}")
+    assert response.status_code == 204
+    response = test_client.delete(f"/connector/{stix_connector_id}")
+    assert response.status_code == 204
+
 
 def test_scheduled_workflow(test_client, caplog, clear_test_suite):
     """
@@ -225,8 +146,11 @@ def test_scheduled_workflow(test_client, caplog, clear_test_suite):
     WHEN the '/login' page is requested (GET)
     THEN check the response is valid
     """
-    stix_config_id = create_connector_and_config_stix(test_client)
-    external_import_config_id = create_connector_and_config_external_import(test_client)
+    (
+        external_import_config_id,
+        external_import_connector_id,
+    ) = create_connector_and_config_external_import(test_client)
+    stix_config_id, stix_connector_id = create_connector_and_config_stix(test_client)
     workflow_id = create_scheduled_workflow(
         [stix_config_id, external_import_config_id], test_client
     )
@@ -247,7 +171,9 @@ def test_scheduled_workflow(test_client, caplog, clear_test_suite):
     run_id = None
     for record in caplog.records:
         if "Received RunContainer" in record.message:
-            run_container = RunContainer(**json.loads(record.message.split("RunContainer: ")[1]))
+            run_container = RunContainer(
+                **json.loads(record.message.split("RunContainer: ")[1])
+            )
             run_id = run_container.run_id
 
     assert run_id is not None, "Could not find Run ID in log output"
@@ -280,3 +206,13 @@ def test_scheduled_workflow(test_client, caplog, clear_test_suite):
 
     assert run.status == State.finished.value
     assert run.result == Result.success.value
+
+    response = test_client.delete(f"/config/{external_import_config_id}")
+    assert response.status_code == 204
+    response = test_client.delete(f"/config/{stix_config_id}")
+    assert response.status_code == 204
+
+    response = test_client.delete(f"/connector/{external_import_connector_id}")
+    assert response.status_code == 204
+    response = test_client.delete(f"/connector/{stix_connector_id}")
+    assert response.status_code == 204
